+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 #### 1.6 (UNRELEASED)
 
 New features and improvements:
-* Added `squiggles` command to apply "shaky hand" style to the geometries (#XXX)  
+* Added `squiggles` command to apply "shaky hand" style to the geometries (#217)  
 
 Other changes:
 * Dropped support for Python 3.6 (#207)